requirements.txt
torch
diffusers
transformers
opencv-python
numpy
Pillow
gradio
